move test to e2e
