+++ test/e2e/app-dir/import-trace-dynamic-eval/lib/foo.ts
@@ -1,0 +1,5 @@
+import { bar } from 'foo'
+
+export function foo() {
+  bar()
+}
