--- test/production/app-dir/import-trace-dynamic-eval/lib/foo.ts
@@ -1,5 +1,0 @@
-import Ajv from 'ajv'
-
-export function foo() {
-  new Ajv().compile({ type: 'string' })
-}
