--- test/production/app-dir/import-trace-dynamic-eval/next.config.js
@@ -1,6 +1,0 @@
-/**
- * @type {import('next').NextConfig}
- */
-const nextConfig = {}
-
-module.exports = nextConfig
